refactor(lab): split offload.rs god-file (#6189)
