+++ homeboy.json
@@ -795,9 +795,8 @@
         "structural::src/core/preview_ingress.rs::HighItemCount",
         "structural::src/core/refactor/rename/mod.rs::GodFile",
         "structural::src/core/runner/connection.rs::HighItemCount",
-        "structural::src/core/runner/lab/agent_task_bridge.rs::HighItemCount",
-        "structural::src/core/runner/lab/offload.rs::GodFile",
-        "structural::src/core/runner/lab_args/tests.rs::HighItemCount",
+         "structural::src/core/runner/lab/agent_task_bridge.rs::HighItemCount",
+         "structural::src/core/runner/lab_args/tests.rs::HighItemCount",
         "structural::src/core/runner/session.rs::HighItemCount",
         "structural::src/core/runner/workspace.rs::GodFile",
         "structural::src/core/runner/workspace.rs::HighItemCount",
