+++ src/core/runner/lab/offload/metadata.rs
@@ -1,0 +1,232 @@
+//! Lab runner / source-checkout / materialization-proof metadata builders,
+//! rig component path overrides, and the stale-runner-homeboy error.
+
+use super::*;
+
+/// Insert generic `${components.<id>.path}` override env vars so a remote rig
+/// check resolves component paths to the runner-side materialized checkout
+/// instead of the controller path the rig spec declares (issue #3766/#3767).
+pub(crate) fn apply_rig_component_path_overrides(
+    env: &mut std::collections::HashMap<String, String>,
+    overrides: &[(String, String)],
+) {
+    for (name, value) in overrides {
+        if !value.trim().is_empty() {
+            env.insert(name.clone(), value.clone());
+        }
+    }
+}
+
+/// Build diagnostics describing each rig component path override forwarded to
+/// the runner, so bench artifacts show how `${components.<id>.path}` resolved.
+pub(crate) fn rig_component_path_overrides_metadata(
+    overrides: &[(String, String)],
+) -> serde_json::Value {
+    let forwarded = overrides
+        .iter()
+        .map(|(name, runner_path)| {
+            serde_json::json!({
+                "env_name": name,
+                "runner_path": runner_path,
+                "forwarded_to_runner": true,
+            })
+        })
+        .collect::<Vec<_>>();
+    serde_json::json!({
+        "schema": "homeboy/lab-offload-rig-component-path-override/v1",
+        "overrides": forwarded,
+    })
+}
+
+pub(crate) fn lab_runner_homeboy_metadata(
+    runner_id: &str,
+    configured_executable: &str,
+    status: &RunnerStatusReport,
+) -> serde_json::Value {
+    let refresh_commands = vec![
+        format!("homeboy runner disconnect {}", shell::quote_arg(runner_id)),
+        format!("homeboy runner connect {}", shell::quote_arg(runner_id)),
+    ];
+    serde_json::json!({
+        "schema": "homeboy/lab-runner-homeboy/v1",
+        "runner_id": runner_id,
+        "configured_executable": configured_executable,
+        "active_daemon_version": status.session.as_ref().map(|session| session.homeboy_version.clone()),
+        "active_daemon_build_identity": status.session.as_ref().and_then(|session| session.homeboy_build_identity.clone()),
+        "stale_daemon": status.stale_daemon,
+        "refresh_commands": refresh_commands,
+        "upgrade_command": format!("homeboy upgrade --force --upgrade-runner {}", shell::quote_arg(runner_id)),
+    })
+}
+
+pub(crate) fn lab_source_checkout_metadata(source_path: &Path) -> serde_json::Value {
+    let git_branch =
+        super::super::super::workspace::git_output(source_path, &["branch", "--show-current"])
+            .ok()
+            .filter(|value| !value.is_empty())
+            .or_else(|| {
+                super::super::super::workspace::git_output(
+                    source_path,
+                    &["rev-parse", "--abbrev-ref", "HEAD"],
+                )
+                .ok()
+            });
+    let git_sha = super::super::super::workspace::git_output(source_path, &["rev-parse", "HEAD"])
+        .ok()
+        .filter(|value| !value.is_empty());
+    let git_remote = super::super::super::workspace::git_output(
+        source_path,
+        &["config", "--get", "remote.origin.url"],
+    )
+    .ok()
+    .filter(|value| !value.is_empty());
+    let dirty =
+        super::super::super::workspace::git_output(source_path, &["status", "--porcelain=v1"])
+            .ok()
+            .map(|status| !status.is_empty());
+
+    serde_json::json!({
+        "schema": "homeboy/lab-source-checkout/v1",
+        "local_path": source_path.display().to_string(),
+        "git_branch": git_branch,
+        "git_sha": git_sha,
+        "git_remote": git_remote,
+        "dirty": dirty,
+    })
+}
+
+pub(crate) fn lab_materialization_proof_metadata(
+    source_snapshot: &SourceSnapshot,
+    workspace_snapshot_identity: &str,
+    remote_workspace: &str,
+    runner_homeboy: &serde_json::Value,
+    source_checkout: &serde_json::Value,
+    workspace_mapping: &serde_json::Value,
+    synced_rigs: &[rig_materialization::LabOffloadRigSync],
+) -> serde_json::Value {
+    serde_json::json!({
+        "schema": "homeboy/lab-materialization-proof/v1",
+        "remote_workspace": remote_workspace,
+        "workload_hashes": {
+            "source_snapshot_hash": source_snapshot.snapshot_hash,
+            "workspace_snapshot_identity": workspace_snapshot_identity,
+        },
+        "source_snapshot": source_snapshot,
+        "source_checkout": source_checkout,
+        "runner_homeboy": runner_homeboy,
+        "workspace_mapping": workspace_mapping,
+        "rigs": synced_rigs,
+    })
+}
+
+pub(crate) fn source_checkout_ref_display(metadata: &serde_json::Value) -> String {
+    let branch = metadata
+        .get("git_branch")
+        .and_then(|value| value.as_str())
+        .filter(|value| !value.is_empty());
+    let sha = metadata
+        .get("git_sha")
+        .and_then(|value| value.as_str())
+        .filter(|value| !value.is_empty())
+        .map(|value| value.chars().take(12).collect::<String>());
+    let dirty = metadata
+        .get("dirty")
+        .and_then(|value| value.as_bool())
+        .map(|value| if value { " dirty" } else { " clean" })
+        .unwrap_or("");
+
+    match (branch, sha) {
+        (Some(branch), Some(sha)) => format!("{branch}@{sha}{dirty}"),
+        (Some(branch), None) => format!("{branch}{dirty}"),
+        (None, Some(sha)) => format!("{sha}{dirty}"),
+        (None, None) => format!("unknown ref{dirty}"),
+    }
+}
+
+pub(crate) fn stale_runner_homeboy_error(
+    runner_id: &str,
+    configured_executable: &str,
+    status: &RunnerStatusReport,
+) -> Error {
+    let refresh_commands = runner_homeboy_refresh_commands(runner_id, status);
+    let active_daemon = status
+        .session
+        .as_ref()
+        .map(runner_session_homeboy_display)
+        .unwrap_or_else(|| "<not connected>".to_string());
+    let current_homeboy = status.stale_daemon.as_ref().map_or_else(
+        || "configured runner executable".to_string(),
+        runner_stale_daemon_current_display,
+    );
+    let drift_message = status
+        .stale_daemon
+        .as_ref()
+        .map(|warning| warning.message.clone())
+        .unwrap_or_else(|| {
+            "connected runner daemon was started by a different Homeboy runtime".to_string()
+        });
+    let refresh = refresh_commands.join(" && ");
+    Error::validation_invalid_argument(
+        "runner",
+        format!(
+            "Lab offload refused runner `{runner_id}` because its active daemon Homeboy/runtime differs from the configured runner executable `{configured_executable}`. Active daemon: {active_daemon}; configured runtime: {current_homeboy}. {drift_message} Stale runner runtimes can return malformed or misleading provider output; reconnect the runner before retrying."
+        ),
+        Some(runner_id.to_string()),
+        Some(vec![
+            format!("Reconnect runner `{runner_id}` before retrying Lab offload: {refresh}"),
+            format!("If the runner binary itself is stale, upgrade it with `homeboy upgrade --force --upgrade-runner {}`.", shell::quote_arg(runner_id)),
+            "Use --force-hot --allow-local-hot only if you intentionally want to bypass Lab offload and run locally.".to_string(),
+        ]),
+    )
+}
+
+pub(crate) fn runner_homeboy_refresh_commands(
+    runner_id: &str,
+    status: &RunnerStatusReport,
+) -> Vec<String> {
+    let commands = status
+        .stale_daemon
+        .as_ref()
+        .map(|warning| warning.recovery_commands.clone())
+        .unwrap_or_default();
+    if !commands.is_empty() && !runner_id.contains(char::is_whitespace) {
+        return commands;
+    }
+    vec![
+        format!("homeboy runner disconnect {}", shell::quote_arg(runner_id)),
+        format!("homeboy runner connect {}", shell::quote_arg(runner_id)),
+    ]
+}
+
+pub(crate) fn runner_session_homeboy_display(
+    session: &super::super::super::RunnerSession,
+) -> String {
+    session
+        .homeboy_build_identity
+        .as_deref()
+        .unwrap_or(&session.homeboy_version)
+        .to_string()
+}
+
+pub(crate) fn runner_stale_daemon_current_display(
+    warning: &super::super::super::RunnerStaleDaemonWarning,
+) -> String {
+    warning
+        .current_homeboy_build_identity
+        .as_deref()
+        .unwrap_or(&warning.current_homeboy_version)
+        .to_string()
+}
+
+pub(crate) fn runner_homeboy_daemon_display(metadata: &serde_json::Value) -> String {
+    metadata
+        .get("active_daemon_build_identity")
+        .and_then(|value| value.as_str())
+        .or_else(|| {
+            metadata
+                .get("active_daemon_version")
+                .and_then(|value| value.as_str())
+        })
+        .unwrap_or("<not connected>")
+        .to_string()
+}
